+++ resources/[dg]/dg-admin/typescript/src/client/controllers/exports.ts
@@ -3,3 +3,7 @@ import { getCmdState } from 'modules/commands/state.commands';
 global.exports('inNoclip', () => {
   return getCmdState('noclip');
 });
+
+global.exports('inCloak', () => {
+  return getCmdState('cloak');
+});