+++ CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Meer data bij confiscate log toegevoegd
 - Kans op uitvliegen met normale gordel verlaagd
 - Kans op voertuig stall bij crash verhoogd
+- Anticheat speedchecks runnen nietmeer als je in admin cloak bent
 
 ### Fixed
 - Je kan weer smelten als je meer dan de nodige items opzak hebt